backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
psycopg2-binary==2.9.5
pydantic==2.8.2
pydantic_core==2.20.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
boto3==1.34.153
python-dotenv==1.0.1
